Refactor: move customselector models to customselector/models
