+++ app/src/main/java/fr/free/nrw/commons/customselector/models/CallbackStatus.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.customselector
+package fr.free.nrw.commons.customselector.models
 
 /**
  * sealed class Callback Status.
