+++ app/src/main/java/fr/free/nrw/commons/customselector/models/UploadedStatus.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.customselector
+package fr.free.nrw.commons.customselector.models
 
 import androidx.room.Entity
 import androidx.room.Index
