+++ app/src/main/java/fr/free/nrw/commons/customselector/ui/selector/ImageFragment.kt
@@ -12,9 +12,9 @@ import androidx.recyclerview.widget.RecyclerView
 import fr.free.nrw.commons.R
 import fr.free.nrw.commons.customselector.helper.ImageHelper
 import fr.free.nrw.commons.customselector.listeners.ImageSelectListener
-import fr.free.nrw.commons.data.models.customselector.CallbackStatus
-import fr.free.nrw.commons.data.models.customselector.Image
-import fr.free.nrw.commons.data.models.customselector.Result
+import fr.free.nrw.commons.customselector.models.CallbackStatus
+import fr.free.nrw.commons.customselector.models.Image
+import fr.free.nrw.commons.customselector.models.Result
 import fr.free.nrw.commons.customselector.ui.adapter.ImageAdapter
 import fr.free.nrw.commons.di.CommonsDaggerSupportFragment
 import fr.free.nrw.commons.theme.BaseActivity
